Scoretize_backend-dev/scoretize_backend/requirements.txt
asgiref==3.5.2
attrs==22.2.0
azure-common==1.1.28
azure-core==1.26.1
azure-identity==1.11.0
azure-mgmt-core==1.3.2
azure-mgmt-datafactory==2.9.0
azure-mgmt-resource==21.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.0.8
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.25.1
drf-yasg==1.21.4
flake8==5.0.4
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.1
mccabe==0.7.0
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
mssql-django==1.1.3
numpy==1.23.4
oauthlib==3.2.2
openai==0.27.7
openapi-codec==1.3.2
packaging==23.0
pandas==1.5.1
pdfkit==1.0.0
Pillow==9.4.0
portalocker==2.6.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyodbc==4.0.34
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
XlsxWriter==3.0.8
azure-storage-blob==12.8.1